backend/requirements_hf.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.24.3

# Database and vector store dependencies
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
qdrant-client==1.8.0
psycopg2-binary==2.9.9
neon==0.2.0

# AI/ML dependencies
openai==1.3.5
cohere>=5.0.0
sentence-transformers==2.7.0
transformers==4.36.0

# Additional dependencies for services
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
unstructured==0.12.5
pypdf==3.17.2
chromadb==0.4.15
sentencepiece==0.1.99

# Utilities
better-exceptions==0.3.3
pytest==7.4.3
pytest-asyncio==0.21.1
tqdm==4.66.1
aiofiles==23.2.1
cachetools==5.3.2

# For text processing
pandas==2.1.0
scikit-learn==1.3.0